services/django/requirements.txt
Django==5.2.4
uvicorn==0.35.0
whitenoise==6.9.0
python-dotenv==1.1.1
pydantic==2.11.7
psycopg[binary,pool]==3.2.9
dj-database-url==3.0.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
